+++ docs/components/filters.mdx
@@ -6,6 +6,7 @@ menu: Components
 import './filters.js'
 import { Playground, PropsTable } from 'docz'
 import Filters from '@emcasa/ui-dom/components/Filters'
+import Button from '@emcasa/ui-dom/components/Filters/FilterButton'
 import SliderRangeFilter from '@emcasa/ui-dom/components/Filters/SliderRangeFilter'
 import ButtonGroupFilter from '@emcasa/ui-dom/components/Filters/ButtonGroupFilter'
 import ButtonRangeFilter from '@emcasa/ui-dom/components/Filters/ButtonRangeFilter'
@@ -16,7 +17,6 @@ import {
   RoomsFilter,
   GarageSpotsFilter
 } from '@emcasa/ui-dom/components/Filters/ListingFilters'
-import Button from '@emcasa/ui-dom/components/Button'
 import View from '@emcasa/ui-dom/components/View'
 
 # Filters
